// src/app/feature/authentication/action/action-type.ts
export enum ActionType {
  LoginRequest = 'LoginRequest',
  LoginSuccess = 'LoginSuccess',
  LoginError = 'LoginError',

  RefreshRequest = 'RefreshRequest',
  RefreshSuccess = 'RefreshSuccess',
  RefreshError = 'RefreshError',

  LogoutRequest = 'LogoutRequest',
  LogoutSuccess = 'LogoutSuccess',
}